[@wtg/redux-modules] Simplify `createModule` internals.
